oops no circular references allowed
it makes node sad :(
